+++ src/Analyzers/CSharp/DiagnosticIdentifiers.Generated.cs
@@ -209,5 +209,6 @@ namespace Roslynator.CSharp
         public const string UseImplicitOrExplicitObjectCreation = "RCS1250";
         public const string RemoveUnnecessaryBraces = "RCS1251";
         public const string NormalizeUsageOfInfiniteLoop = "RCS1252";
+        public const string FormatDocumentationCommentSummary = "RCS1253";
     }
 }