+++ src/Common/ConfigOptionKeys.Generated.cs
@@ -19,6 +19,7 @@ namespace Roslynator
         public const string ConditionalOperatorConditionParenthesesStyle      = "roslynator_conditional_operator_condition_parentheses_style";
         public const string ConditionalOperatorNewLine                        = "roslynator_conditional_operator_new_line";
         public const string ConfigureAwait                                    = "roslynator_configure_await";
+        public const string DocCommentSummaryStyle                            = "roslynator_doc_comment_summary_style";
         public const string EmptyStringStyle                                  = "roslynator_empty_string_style";
         public const string EnumHasFlagStyle                                  = "roslynator_enum_has_flag_style";
         public const string EqualsTokenNewLine                                = "roslynator_equals_token_new_line";