+++ src/VisualStudioCode/package/src/configurationFiles.generated.ts
@@ -72,6 +72,9 @@ roslynator_analyzers.enabled_by_default = true|false
 #roslynator_configure_await = true|false
 # Applicable to: rcs1090
 
+#roslynator_doc_comment_summary_style = multi_line|single_line
+# Applicable to: rcs1253
+
 #roslynator_empty_string_style = field|literal
 # Applicable to: rcs1078
 
@@ -514,12 +517,6 @@ roslynator_analyzers.enabled_by_default = true|false
 # Default label should be the last label in a switch section
 #dotnet_diagnostic.rcs1099.severity = suggestion
 
-# Format documentation summary on a single line
-#dotnet_diagnostic.rcs1100.severity = none
-
-# Format documentation summary on multiple lines
-#dotnet_diagnostic.rcs1101.severity = none
-
 # Make class static
 #dotnet_diagnostic.rcs1102.severity = warning
 
@@ -899,6 +896,10 @@ roslynator_analyzers.enabled_by_default = true|false
 # Normalize usage of infinite loop
 #dotnet_diagnostic.rcs1252.severity = none
 # Options: roslynator_infinite_loop_style
+
+# Format documentation comment summary
+#dotnet_diagnostic.rcs1253.severity = none
+# Options: roslynator_doc_comment_summary_style
 
 # Use pattern matching
 #dotnet_diagnostic.rcs9001.severity = silent